ui: Add CopyToClipboardButton widget (#2799)
A button that copies text to the clipboard when clicked and also
provides some visual feedback.


![copy-to-clipboard-button](https://github.com/user-attachments/assets/a648a5f4-bf77-4d06-95c3-aa2145181bf1)
